requirements.txt
backports.statistics==0.1.0
Cython==0.29.24
deepface==0.0.68
dl==0.1.0
Flask==2.0.2
imageio==2.10.4
ipaddr==2.2.0
matplotlib==3.3.4
numpy==1.19.5
ordereddict==1.1
pandas==1.1.5
Pillow==8.4.0
torch==1.10.0
torchvision==0.11.1
urllib3==1.26.7
